refactor: move data utils to api
