+++ providers/ic/src/providers/data/paragraph.providers.ts
@@ -1,5 +1,5 @@
 import {Paragraph, ParagraphData} from '@deckdeckgo/editor';
-import {getData} from '../../utils/data.utils';
+import {getData} from '../../api/data.api';
 
 export const getParagraph = (docId: string, paragraphId: string): Promise<Paragraph | undefined> =>
   getData<Paragraph, ParagraphData>({
