+++ providers/ic/src/providers/data/slide.providers.ts
@@ -1,5 +1,5 @@
 import {Slide, SlideData} from '@deckdeckgo/editor';
-import {getData} from '../../utils/data.utils';
+import {getData} from '../../api/data.api';
 
 export const getSlide = (deckId: string, slideId: string): Promise<Slide | undefined> =>
   getData<Slide, SlideData>({key: `/decks/${deckId}/slides/${slideId}`});
