+++ providers/ic/src/providers/data/user.providers.ts
@@ -1,5 +1,5 @@
 import {log, UpdateUser, User, UserData} from '@deckdeckgo/editor';
-import {setData} from '../../utils/data.utils';
+import {setData} from '../../api/data.api';
 
 export const updateUser: UpdateUser = async (user: User): Promise<User> => {
   log({msg: '[update][start] user'});
