+++ providers/ic/src/providers/feed/feed.providers.ts
@@ -12,7 +12,7 @@ import {Identity} from '@dfinity/agent';
 import {_SERVICE as FeedActor} from '../../canisters/feed/feed.did';
 import {_SERVICE as StorageBucketActor} from '../../canisters/storage/storage.did';
 import {EnvStore} from '../../stores/env.store';
-import {setData} from '../../utils/data.utils';
+import {setData} from '../../api/data.api';
 import {toNullable} from '../../utils/did.utils';
 import {createFeedActor} from '../../utils/feed.utils';
 import {BucketActor, getStorageBucket} from '../../utils/manager.utils';
