+++ providers/ic/src/utils/publish.deck.utils.ts
@@ -1,6 +1,6 @@
 import {Deck, DeckData, DeckPublishData, deckPublishData, PublishData} from '@deckdeckgo/editor';
 import {EnvStore} from '../stores/env.store';
-import {setData} from './data.utils';
+import {setData} from '../api/data.api';
 import {uploadSocialImage} from './publish.social.utils';
 import {
   initIndexHTML,
